(maint) Fix members property getter to report the right change notification
Previously, the members property getter would return the should value if
auth_membership == false and the current members contained all of the
members specified in the should value. This was done to properly implement
the insync? method, which would (correctly) return true in this situation.
Unfortunately, PUP 5.5.7's refactoring of the members property to derive
from Puppet::Property::List means that the members' property's should
method now uses Puppet::Property::List's should method, the latter of
which depends on the members property getter. Specifically, it expects
the members property getter to always return _all_ of the group's
members. Because the current implementation does not always satisfy this
precondition, Puppet would end up reporting the wrong change
notification in certain situations (e.g. when adding new members with
auth_membership == false and specifying these members as an array).

This commit refactors the members property getter to return _all_ of the
group's members. This in-turn leads to a refactor of the insync? and
members property setter methods.

Note that due to PUP-6542, older versions of Puppet will still report the
wrong change notification even with these changes.